+++ src/async.ts
@@ -345,7 +345,13 @@ type CacheObject<T = any> = {
 /*#__NO_SIDE_EFFECTS__*/
 export const cachifyAsync = <
   R,
-  F extends (...args: Array<any>) => Promise<Result.TupleObj<R>>
+  F extends (...args: Array<any>) => Promise<Result.TupleObj<R>>,
+  UnWrap extends boolean = false,
+  RetF = boolean extends UnWrap
+    ? F
+    : UnWrap extends false
+    ? F
+    : (...args: Parameters<F>) => Promise<Awaited<ReturnType<F>>['result']>
 >(opts: {
   /** The async function to cache. */
   fn: F;
@@ -416,13 +422,30 @@ export const cachifyAsync = <
    * Default: `true`
    */
   returnStale?: boolean;
-}): F & ClearCache<F> => {
+
+  /**
+   * Sugar option to "unwrap" the cached function's return value to only the `result`
+   * property of the `Result.TupleObj` returned by the underlying function.
+   *
+   * In cases where the cached function returns a `Fail` tuple, the `error` is
+   * ignored and the `undefined` `result` value is returned.
+   *
+   * This is useful when the cached function only ever returns "Success" results
+   * and you want to skip the extra `Result.TupleObj` wrapper, or when you only
+   * care about the successful result and want to ignore/swallow any errors and
+   * convert them to `undefined`.
+   *
+   * Default: `false`
+   */
+  unwrapResult?: UnWrap;
+}): RetF & ClearCache<F> => {
   const {
     fn,
     getKey = (...args) => JSON.stringify(args),
     customTtl,
     cache,
     returnStale = true,
+    unwrapResult,
   } = opts;
 
   // Set up the cache object
@@ -442,7 +465,7 @@ export const cachifyAsync = <
     const key = getKey(...args);
     const cached = _cache.get(key);
     if (cached && now < cached.freshUntil) {
-      return cached.data;
+      return unwrapResult ? cached.data.then((data) => data.result) : cached.data;
     }
 
     const msScaling =
@@ -498,8 +521,8 @@ export const cachifyAsync = <
     };
     _cache.set(key, entry);
 
-    return entry.data;
-  }) as F & ClearCache<F>;
+    return unwrapResult ? entry.data.then((data) => data.result) : entry.data;
+  }) as RetF & ClearCache<F>;
 
   cachedFn.invalidate = ((...fnArgs) => {
     const key = getKey(...fnArgs);